requirements.txt
streamlit==1.7.0
Pillow==9.0.1